requirements.txt
flask
flask-cors
pillow
numpy
scipy
torch
torchvision
scikit-learn